Add files via upload
Adding code for making the radiative-capture fraction calculations from 25Mg(p,g)26Al ground state and isomeric reactions. This code is a fairly horrible hacked together thing but the idea is that it reads the input from the ENSDF file and then will cascade the gamma ray transitions down to either the ground state or the isomer. Hopefully (!) if the ENSDF files are updated with new gamma-ray branches then the code will still work and can be used to update the ground-state fraction.